Moved some functions from gui_tools to new utils directory.
